refactor: $TSFixMe修正

+++ src/app/feat-admin/channel_manager/deleteCategory.ts
@@ -1,23 +1,27 @@
 import fs from 'fs';
+import { request } from 'http';
 
 import { parse } from 'csv';
 import { stringify } from 'csv-stringify/sync';
-import { AttachmentBuilder, ChannelType, PermissionsBitField } from 'discord.js';
-import request from 'request';
+import { AttachmentBuilder, ChannelType, ChatInputCommandInteraction, Guild, PermissionsBitField } from 'discord.js';
 
 import { log4js_obj } from '../../../log4js_settings';
 import { searchChannelById } from '../../common/manager/channel_manager';
-import { exists, notExists } from '../../common/others';
+import { getGuildByInteraction } from '../../common/manager/guild_manager';
+import { searchAPIMemberById } from '../../common/manager/member_manager';
+import { assertExistCheck, exists, notExists } from '../../common/others';
 
 const logger = log4js_obj.getLogger('ChannelManager');
 
-export async function handleDeleteCategory(interaction: $TSFixMe) {
-    if (!interaction.inGuild()) return;
-
+export async function handleDeleteCategory(interaction: ChatInputCommandInteraction<'cached' | 'raw'>) {
     // 'インタラクションに失敗'が出ないようにするため
     await interaction.deferReply();
 
-    if (!interaction.member.permissions.has(PermissionsBitField.Flags.ManageChannels)) {
+    const guild = await getGuildByInteraction(interaction);
+    const member = await searchAPIMemberById(guild, interaction.member.user.id);
+    assertExistCheck(member, 'member');
+
+    if (!member.permissions.has(PermissionsBitField.Flags.ManageChannels)) {
         return await interaction.followUp('チャンネルを管理する権限がないでし！');
     }
 
@@ -39,7 +43,7 @@ export async function handleDeleteCategory(interaction: $TSFixMe) {
         await interaction.editReply('CSVを読み込んで削除中でし！\nちょっと待つでし！');
 
         request(attachment.url).pipe(
-            parse(async function (err: $TSFixMe, data: $TSFixMe) {
+            parse(async function (err, data) {
                 let categoryIdList = [];
                 try {
                     for (const i in data) {
@@ -50,21 +54,21 @@ export async function handleDeleteCategory(interaction: $TSFixMe) {
                     logger.error(error);
                     await interaction.followUp('CSVファイル読み込み中にエラーでし！');
                 }
-                deleteCategory(interaction, categoryIdList);
+                await deleteCategory(interaction, categoryIdList);
             }),
         );
     } else if (args.length != 0) {
         await interaction.editReply('指定されたIDのカテゴリを削除中でし！\nちょっと待つでし！');
         const categoryIdList = Array.from(new Set(args));
-        deleteCategory(interaction, categoryIdList);
+        await deleteCategory(interaction, categoryIdList);
     } else {
         await interaction.followUp('CSVファイルを添付するか、削除したいカテゴリのIDを入れるでし！');
         return;
     }
 }
 
-async function deleteCategory(interaction: $TSFixMe, categoryIdList: $TSFixMe) {
-    const guild = await interaction.guild.fetch();
+async function deleteCategory(interaction: ChatInputCommandInteraction<'cached' | 'raw'>, categoryIdList: string[]) {
+    const guild = await (await getGuildByInteraction(interaction)).fetch();
     const removed = [];
 
     removed.push(['カテゴリID', 'カテゴリ名', 'チャンネルID', 'チャンネル名']);
@@ -85,7 +89,12 @@ async function deleteCategory(interaction: $TSFixMe, categoryIdList: $TSFixMe) {
             } else {
                 const channels = await deleteChannelsByCategoryId(guild, categoryId);
                 const channelCollection = await guild.channels.fetch();
-                const category = channelCollection.find((c: $TSFixMe) => c.id == categoryId && c.type == ChannelType.GuildCategory);
+                const category = channelCollection.find(
+                    (channel) => exists(channel) && channel.id == categoryId && channel.type == ChannelType.GuildCategory,
+                );
+                if (notExists(category)) {
+                    continue;
+                }
                 categoryName = category.name;
                 await category.delete();
                 await guild.channels.fetch();
@@ -117,11 +126,20 @@ async function deleteCategory(interaction: $TSFixMe, categoryIdList: $TSFixMe) {
     });
 }
 
-async function deleteChannelsByCategoryId(guild: $TSFixMe, categoryId: $TSFixMe) {
+async function deleteChannelsByCategoryId(guild: Guild, categoryId: string) {
     const channels = [];
     let channelCollection = await guild.channels.fetch();
-    while (exists(channelCollection.find((c: $TSFixMe) => c.type != ChannelType.GuildCategory && c.parent == categoryId))) {
-        const channel = channelCollection.find((c: $TSFixMe) => c.type != ChannelType.GuildCategory && c.parent == categoryId);
+    while (
+        exists(
+            channelCollection.find(
+                (c) => exists(c) && c.type !== ChannelType.GuildCategory && exists(c.parent) && c.parent.id === categoryId,
+            ),
+        )
+    ) {
+        const channel = channelCollection.find(
+            (c) => exists(c) && c.type != ChannelType.GuildCategory && exists(c.parent) && c.parent.id === categoryId,
+        );
+        assertExistCheck(channel, 'channel');
         if (channel.type == ChannelType.GuildText) {
             channels.push([channel.id, '#' + channel.name]);
         } else if (channel.type == ChannelType.GuildVoice) {
